brrrr

+++ src/main/java/org/metamechanists/aircraft/vehicle/forces/AerodynamicCoefficients.java
@@ -1,3 +1,3 @@
 package org.metamechanists.aircraft.vehicle.forces;
 
-public record AerodynamicCoefficients(int dragCoefficient, int liftCoefficient) {}
+public record AerodynamicCoefficients(double dragCoefficient, double liftCoefficient) {}
